Shorten `class enum`s as much as possible
Fixing clazy warnings along the way

* Add a missing declaration
* Add/remove override/virtual
* Remove unused includes, sort the used ones
* Fully qualify types in slots and signals
* And quite a bit more, basicaly a full clazy run

+++ libmscore/scoreElement.h
@@ -135,7 +135,7 @@ class MMRestRange;
 
 class StaffTextBase;
 
-enum class Pid : int;
+enum class Pid : short;
 enum class PropertyFlags : char;
 
 //---------------------------------------------------------
